// src/app/models/metadata.model.ts
export interface ReqMetadata {
    deviceType?: string;
        deviceVersion?: string;
        deviceToken?: string;
        deviceOS?: string;
        publicIp?: string;
        privateIp?: string;
        logitude?: string;
        latitude?: string;
        appVersion?: string;
        address?: string;
        city?: string;
        state?: string;
        country?: string;
        zipCode?: string;
        accessPlateform?: string;
        carrier?: string;
        userAgent?: string;
        brand?: string;
        model?: string;
}